fix: collect NOAA articles via playwright
