+++ tests/spiders/test_noaa.py
@@ -1,9 +1,6 @@
-from typing import Any
+from scrapy.http import HtmlResponse
 
-import pytest
-
-from open_ire.errors import SpiderParameterError
-from open_ire.settings import OPEN_IRE_DEFAULT_TERMS
+from open_ire.settings import OPEN_IRE_SEARCH_TERMS
 from open_ire.spiders.noaa import NOAASpider
 
 
@@ -12,177 +9,163 @@ class TestNOAASpider:
         """Test spider initialization with default parameters."""
         spider = NOAASpider()
         assert spider.name == "noaa"
-        assert spider.terms == OPEN_IRE_DEFAULT_TERMS.lower().split(",")
+        assert len(spider.start_urls) == len(OPEN_IRE_SEARCH_TERMS)
+        assert "repository.library.noaa.gov" in spider.start_urls[0]
+        assert "maxResults=100" in spider.start_urls[0]
+        assert "start=" not in spider.start_urls[0]
 
     def test_custom_params(self):
         """Test spider initialization with custom parameters."""
-        terms = "unittest,test,sample"
-        spider = NOAASpider(terms=terms)
-        assert spider.terms == ["unittest", "test", "sample"]
+        terms = "climate,ocean"
+        page = "2"
+        spider = NOAASpider(terms=terms, page=page)
 
-    def test_page_param_not_supported(self):
-        """Test that page parameter raises ValueError."""
-        with pytest.raises(SpiderParameterError) as e:
-            NOAASpider(page="2")
+        assert spider.target_page == 2
+        assert len(spider.start_urls) == 2
+        assert "terms=climate" in spider.start_urls[0]
+        assert "terms=ocean" in spider.start_urls[1]
+        assert "maxResults=100" in spider.start_urls[0]
+        assert "start=100" in spider.start_urls[0]
 
-        assert e.value.parameter == "page"
-        assert e.value.spider_name == NOAASpider.name
-
-
-    def test_normalize_terms(self):
-        """Test term normalization."""
-        terms = "Unittest, Test , Sample,  "
-        normalized = NOAASpider._normalize_terms(terms)
-        assert normalized == ["unittest", "test", "sample"]
-
-    def test_get_field_value(self):
-        """Test field value extraction."""
-        doc = {
-            "field1": "value1",
-            "field2": ["value2a", "value2b"],
-            "field3": None,
-        }
-
-        assert NOAASpider._get_field_value(doc, ["missing", "field1"]) == "value1"
-        assert NOAASpider._get_field_value(doc, ["field2"]) == ['value2a', 'value2b']
-        assert NOAASpider._get_field_value(doc, ["missing"]) is None
-
-    def test_normalize_field_value(self):
-        """Test field value normalization."""
-        assert NOAASpider._normalize_field_value("unittest") == "unittest"
-        assert NOAASpider._normalize_field_value(["first", "second"]) == "first"
-        assert NOAASpider._normalize_field_value([]) is None
-        assert NOAASpider._normalize_field_value(None) is None
-        assert NOAASpider._normalize_field_value("  spaced  ") == "spaced"
-
-    def test_extract_authors(self):
-        """Test author extraction."""
+    def test_parse(self):
+        """Test the parse method extracts article links and next page."""
+        html = """
+        <html>
+            <body>
+                <div class="search-result-row">
+                    <div class="object-title">
+                        <a href="/view/noaa/123">Article 1</a>
+                    </div>
+                </div>
+                <div class="search-result-row">
+                    <div class="object-title">
+                        <a href="/view/noaa/456">Article 2</a>
+                    </div>
+                </div>
+                <a id="nextPage" href="/gsearch?start=100">Next</a>
+            </body>
+        </html>
+        """
+        response = HtmlResponse(
+            url="https://repository.library.noaa.gov/gsearch", body=html.encode("utf-8")
+        )
         spider = NOAASpider()
 
-        # Single author
-        doc: dict[str, Any] = {"mods.name_personal": "Unittest Author"}
-        assert spider._extract_authors(doc) == "Unittest Author"
+        requests = list(spider.parse(response))
 
-        # Multiple authors
-        doc = {"mods.name_personal": ["Unittest Author 1", "Unittest Author 2"]}
-        assert spider._extract_authors(doc) == "Unittest Author 1, Unittest Author 2"
+        assert len(requests) == 3  # 2 articles + 1 next page
+        assert requests[0].callback == spider.parse_detail
+        assert requests[0].url == "https://repository.library.noaa.gov/view/noaa/123"
+        assert requests[1].url == "https://repository.library.noaa.gov/view/noaa/456"
+        assert (
+            requests[2].url == "https://repository.library.noaa.gov/gsearch?start=100"
+        )
 
-        # No authors
-        doc = {}
-        assert spider._extract_authors(doc) is None
-
-    def test_extract_publication_date(self):
-        """Test publication date extraction."""
+    def test_parse_detail(self):
+        """Test the parse_detail method extracts article metadata correctly."""
+        html = """
+        <html>
+            <head>
+                <meta name="citation_title" content="Unit Test Title">
+                <meta name="citation_abstract" content="Unit Test Abstract">
+                <meta name="citation_doi" content="10.1000/1234">
+                <meta name="citation_author" content="Unit Test Author 1">
+                <meta name="citation_author" content="Unit Test Author 2">
+                <meta name="citation_publication_date" content="2025-07-25">
+                <meta name="citation_issn" content="1234-5678">
+                <meta name="citation_pdf_url" content="/documents/test.pdf">
+                <meta name="citation_volume" content="Volume !">
+            </head>
+            <body></body>
+        </html>
+        """
+        response = HtmlResponse(
+            url="https://repository.library.noaa.gov/view/noaa/1234/",
+            body=html.encode("utf-8"),
+        )
         spider = NOAASpider()
 
-        # Valid date
-        doc = {"mods.ss_publishyear": "2025"}
-        result = spider._extract_publication_date(doc)
-        assert result is not None
-        assert result.year == 2025
+        items = list(spider.parse_detail(response))
 
-        # Invalid date
-        doc = {"mods.ss_publishyear": "invalid"}
-        assert spider._extract_publication_date(doc) is None
+        assert len(items) == 1
+        item = items[0]
+        assert item.title == "Unit Test Title"
+        assert item.abstract == "Unit Test Abstract"
+        assert item.doi == "10.1000/1234"
+        assert item.authors == "Unit Test Author 1, Unit Test Author 2"
+        assert str(item.publication_date) == "2025-07-25"
+        assert item.issn == "1234-5678"
+        assert item.reference == "1234"
+        assert item.repository == "noaa"
+        assert item.url == "https://repository.library.noaa.gov/view/noaa/1234/"
+        assert len(item.file_urls) == 1
+        assert hasattr(item, "extra")
 
-        # No date
-        doc = {}
-        assert spider._extract_publication_date(doc) is None
+    def test_extract_file_urls(self):
+        """Test the extract_file_urls method."""
+        html = """
+        <html>
+            <head>
+                <meta name="citation_pdf_url" content="/documents/file1.pdf">
+                <meta name="citation_pdf_url" content="/documents/file2.pdf">
+            </head>
+            <body></body>
+        </html>
+        """
+        response = HtmlResponse(
+            url="https://repository.library.noaa.gov/", body=html.encode("utf-8")
+        )
 
-    def test_extract_keywords(self):
-        """Test keyword extraction."""
+        urls = NOAASpider._extract_file_urls(response)
 
-        keywords = NOAASpider._extract_keywords(["unittest", "test", "unittest"])
-        assert len(keywords) == 2
-        assert "unittest" in keywords
-        assert "test" in keywords
-
-        # Single keyword
-        keywords = NOAASpider._extract_keywords("sample")
-        assert keywords == ["sample"]
-
-        # Empty
-        keywords = NOAASpider._extract_keywords("")
-        assert keywords == []
+        assert len(urls) == 2
+        assert urls[0] == "https://repository.library.noaa.gov/documents/file1.pdf"
+        assert urls[1] == "https://repository.library.noaa.gov/documents/file2.pdf"
 
     def test_extract_extra_details(self):
-        """Test extra details extraction."""
-        spider = NOAASpider()
-        doc = {
-            "mods.sm_publisher": "Unittest Publications",
-            "mods.volume": "Volume 1",
-            "mods.subject_topic": ["unittest", "test"],
-        }
+        """Test the extract_extra_details method."""
+        html = """
+        <html>
+            <head>
+                <meta name="citation_volume" content="Volume 1">
+                <meta name="citation_publisher" content="NOAA Publications">
+                <meta name="citation_journal_title" content="Unit Test Journal">
+                <meta name="citation_conference" content="Unit Test Conference">
+                <meta name="citation_language" content="English">
+                <meta name="citation_keywords" content="climate">
+                <meta name="citation_keywords" content="ocean">
+                <meta name="citation_keywords" content="climate">
+            </head>
+            <body>
+                <textarea id="Genericpreview">Unit Test Citation</textarea>
+            </body>
+        </html>
+        """
+        response = HtmlResponse(
+            url="https://repository.library.noaa.gov/", body=html.encode("utf-8")
+        )
 
-        extra = spider._extract_extra_details(doc)
-        assert extra["publisher"] == "Unittest Publications"
+        spider = NOAASpider()
+        extra = spider._extract_extra_details(response)
+
         assert extra["volume"] == "Volume 1"
-        assert "unittest" in extra["keywords"]
-        assert "test" in extra["keywords"]
+        assert extra["publisher"] == "NOAA Publications"
+        assert extra["journal_title"] == "Unit Test Journal"
+        assert extra["conference"] == "Unit Test Conference"
+        assert extra["language"] == "English"
+        assert extra["citation_text"] == "Unit Test Citation"
+        assert "climate" in extra["keywords"]
+        assert "ocean" in extra["keywords"]
+        assert len(extra["keywords"]) == 2
 
-    def test_document_matches_terms(self):
-        """Test document term matching."""
-        spider = NOAASpider(terms="unittest,test")
+    def test_extract_extra_details_empty(self):
+        """Test the extract_extra_details method with an empty response."""
+        html = "<html><head></head><body></body></html>"
+        response = HtmlResponse(
+            url="https://repository.library.noaa.gov/", body=html.encode("utf-8")
+        )
 
-        # Matching document
-        doc = {
-            "mods.title": "Unittest Sample Study",
-            "mods.abstract": "This study examines unittest patterns.",
-        }
-        assert spider._document_matches_terms(doc)
-
-        # Non-matching document
-        doc = {
-            "mods.title": "Sample Study",
-            "mods.abstract": "This study examines sample practices.",
-        }
-        assert not spider._document_matches_terms(doc)
-
-        # No terms (should match all)
-        spider_no_terms = NOAASpider(terms="")
-        assert spider_no_terms._document_matches_terms(doc)
-
-    def test_build_searchable_text(self):
-        """Test searchable text building."""
         spider = NOAASpider()
-        doc = {
-            "mods.title": "Unittest Study",
-            "mods.abstract": "Test Sample Analysis",
-            "mods.name_personal": ["Unittest Author"],
-            "missing_field": "Should not be included",
-        }
+        extra = spider._extract_extra_details(response)
 
-        searchable = spider._build_searchable_text(doc)
-        assert "unittest study" in searchable
-        assert "test sample analysis" in searchable
-        assert "unittest author" in searchable
-        assert "should not be included" not in searchable
-
-    def test_create_article_item(self):
-        """Test article item creation."""
-        spider = NOAASpider()
-        doc = {
-            "PID": "noaa:12345",
-            "mods.title": "Unittest Article",
-            "mods.abstract": "Unittest abstract",
-            "mods.sm_digital_object_identifier": "10.1000/unittest",
-            "mods.name_personal": "Unittest Author",
-            "mods.ss_publishyear": "2025",
-            "mods.sm_issn": "1234-5678",
-            "mods.sm_publisher": "Unittest Publisher",
-        }
-
-        item = spider._create_article_item(doc)
-
-        assert item.reference == "12345"
-        assert item.title == "Unittest Article"
-        assert item.abstract == "Unittest abstract"
-        assert item.doi == "10.1000/unittest"
-        assert item.authors == "Unittest Author"
-        assert item.publication_date.year == 2025
-        assert item.issn == "1234-5678"
-        assert item.repository == "noaa"
-        assert item.url == "https://repository.library.noaa.gov/view/noaa/12345"
-        assert len(item.file_urls) == 1
-        assert "noaa_12345_DS1.pdf" in item.file_urls[0]
-        assert item.extra["publisher"] == "Unittest Publisher"
+        assert extra == {}
